requirements.txt
requests
python-dotenv
tqdm

# For --local mode (sentence-transformers)
torch
sentence-transformers
